abstract out some stuff
