+++ server/src/server.js
@@ -4,6 +4,7 @@ const helmet = require('helmet');
 const api = require('./api');
 const logger = require('./utils/logger');
 const checkJwt = require('./utils/auth');
+const constants = require('./utils/stringConstants');
 
 const app = express();
 
@@ -24,13 +25,13 @@ app.use('/api/v1', api);
 app.use((err, req, res, next) => {
   switch (err.name) {
     case 'UnauthorizedError':
-      logger.error('Request is not authorized.');
-      res.status(401).json('Access... DENIED.');
+      logger.error(constants.errors.UNAUTHORIZED);
+      res.status(401).json(constants.errors.UNAUTHORIZED);
       break;
 
     default:
       logger.error(`${err.name}: ${err.message}`);
-      res.status(500).json('Something broke!');
+      res.status(500).json(`${err.name}: ${err.message}`);
   }
 });
