+++ server/src/test/integration/drinksEndpointTest.js
@@ -2,8 +2,9 @@ const app = require('../../server');
 const request = require('supertest');
 const expect = require('chai').expect;
 const dbAdapter = require('../../db/adapter');
-const auth = require('../../../config').auth;
 const axios = require('axios');
+const testObjects = require('../../utils/testObjects');
+const constants = require('../../utils/stringConstants');
 
 describe(`'drinks' route - api test`, function() {
   let token;
@@ -11,31 +12,14 @@ describe(`'drinks' route - api test`, function() {
   before(async function() {
     // prime the database with test tables/data
     const tables = await dbAdapter.r.tableList();
-    const drinkTestObjects = [
-      {
-        ingredients: ['whiskey', 'sugar', 'bitters'],
-        name: 'Old Fashioned'
-      },
-      {
-        ingredients: ['gin', 'sugar', 'lemon juice', 'club soda'],
-        name: 'Tom Collins'
-      }
-    ];
+
     if (tables.includes('drinks')) {
       await dbAdapter.r.tableDrop('drinks');
     }
     await dbAdapter.r.tableCreate('drinks', { primaryKey: 'name' });
-    await dbAdapter.r.table('drinks').insert(drinkTestObjects);
+    await dbAdapter.r.table('drinks').insert(testObjects.drinkTest);
 
-    // get a temporary bearer token for testing
-    const options = {
-      method: 'POST',
-      url: 'https://bartop.auth0.com/oauth/token',
-      headers: { 'content-type': 'application/json' },
-      data: `{"client_id":"${auth.id}","client_secret":"${auth.secret}","audience":"${auth.audience}","grant_type":"${auth.grant}"}`
-    };
-
-    const response = await axios(options);
+    const response = await axios(testObjects.tokenRequestOptions);
     token = response.data.access_token;
 
     return;
@@ -59,7 +43,7 @@ describe(`'drinks' route - api test`, function() {
       .get('/api/v1/drinks')
       .end((err, res) => {
         expect(res.statusCode).to.equal(401);
-        expect(res.body).to.equal('Access... DENIED.');
+        expect(res.body).to.equal(constants.errors.UNAUTHORIZED);
         done();
       });
   });
@@ -71,6 +55,6 @@ describe(`'drinks' route - api test`, function() {
       .set('Authorization', 'Bearer ' + token);
 
     expect(res.statusCode).to.equal(500);
-    expect(res.body).to.equal('Something broke!');
+    expect(res.body.split(':')[0]).to.equal('ReqlOpFailedError');
   });
 });
