+++ server/src/test/integration/usersEndpointTest.js
@@ -2,8 +2,9 @@ const app = require('../../server');
 const request = require('supertest');
 const expect = require('chai').expect;
 const dbAdapter = require('../../db/adapter');
-const auth = require('../../../config').auth;
 const axios = require('axios');
+const options = require('../../utils/testObjects').tokenRequestOptions;
+const constants = require('../../utils/stringConstants');
 
 describe(`'users' route - api test`, function() {
   let token;
@@ -17,13 +18,6 @@ describe(`'users' route - api test`, function() {
     await dbAdapter.r.tableCreate('users');
 
     // get a temporary bearer token for testing
-    const options = {
-      method: 'POST',
-      url: 'https://bartop.auth0.com/oauth/token',
-      headers: { 'content-type': 'application/json' },
-      data: `{"client_id":"${auth.id}","client_secret":"${auth.secret}","audience":"${auth.audience}","grant_type":"${auth.grant}"}`
-    };
-
     const response = await axios(options);
     token = response.data.access_token;
 
@@ -54,7 +48,7 @@ describe(`'users' route - api test`, function() {
       .post('/api/v1/users/12345')
       .end((err, res) => {
         expect(res.statusCode).to.equal(401);
-        expect(res.body).to.equal('Access... DENIED.');
+        expect(res.body).to.equal(constants.errors.UNAUTHORIZED);
         done();
       });
   });
@@ -66,6 +60,6 @@ describe(`'users' route - api test`, function() {
       .set('Authorization', 'Bearer ' + token);
 
     expect(res.statusCode).to.equal(500);
-    expect(res.body).to.equal('Something broke!');
+    expect(res.body.split(':')[0]).to.equal('ReqlOpFailedError');
   });
 });
