+++ server/src/utils/stringConstants.js
@@ -1,0 +1,5 @@
+module.exports = {
+  errors: {
+    UNAUTHORIZED: 'Request is not authorized.'
+  }
+};
